requirements.txt
pandas
Flask==1.1.2
celery==4.1.1
SQLAlchemy==1.4.39
oauth2client==4.1.3
oauthlib==3.2.0
psycopg2==2.9.3
Flask-SQLAlchemy
google-api-core==2.8.2
python-dotenv
google-api-python-client==2.55.0
google-auth==2.9.1
google-auth-httplib2==0.1.0
markupsafe==2.0.1
google-auth-oauthlib==0.5.2
python-cbr
gspread
Jinja2==2.10.2
itsdangerous==2.0.1
werkzeug==2.0.3
tasks
Flask-APScheduler